fix(manifest): stats should add prefetchInterface if enable dataPrefetch
