+++ .changeset/shaggy-ligers-teach.md
@@ -1,0 +1,5 @@
+---
+'@module-federation/manifest': patch
+---
+
+fix(manifest): stats should add prefetchInterface if enable dataPrefetch
