+++ packages/manifest/src/ManifestManager.ts
@@ -1,4 +1,3 @@
-import path from 'path';
 import fs from 'fs';
 import chalk from 'chalk';
 import {
@@ -102,20 +101,6 @@ class ManifestManager {
       sum.push(remote);
       return sum;
     }, [] as ManifestRemote[]);
-
-    let prefetchInterface = false;
-    const prefetchFilePath = path.resolve(
-      compiler.options.context || process.cwd(),
-      `node_modules/.mf/${encodeName(stats.name)}/${MFPrefetchCommon.fileName}`,
-    );
-    const existPrefetch = fs.existsSync(prefetchFilePath);
-    if (existPrefetch) {
-      const content = fs.readFileSync(prefetchFilePath).toString();
-      if (content) {
-        prefetchInterface = true;
-      }
-    }
-    stats.metaData.prefetchInterface = prefetchInterface;
 
     this._manifest = manifest;
